less mentions of Silk Road
The concept uses Silk Road as the model of an anonymous marketplaces.
Due to the nature of things that were primarily traded on SR, some
people get scared away by encountering "Silk Road" in the title.
This concept, however, is an attempt to describe a better structure of
organizing anonymous marketplaces, and not an attempt to promote any
specific use of such a marketplace. Because of these reasons, it's
better to give a more neutral name to this concept.